pyproject.toml
[tool.poetry]
name = "notion-calendar"
version = "0.1.0"
description = ""
authors = ["Jonathan Breidfjord <jbreidfjord@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
typer = { extras = ["all"], version = "^0.6.1" }

[tool.poetry.dev-dependencies]
black = "^22.8.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
